いったん普通のfix (#8)
* fix: 軽微なもの

* docs: update changelog with recent project changes

* feat: add auto-generated definition files for JMA and CAMF data structures and include NankaiPageBuffer

* fix: convert svid from UBX-based GNSS module

* fix: fixed svid at testcase

* feat: implement UbxFramer and add comprehensive test suite for framer modules

* feat: implement QZSS DCX and QZQSM message parsing, aggregation, and serialization support

* feat: implement DCX decoder, JSON serializers, and comprehensive test suite for message processing

* feat: implement JSON serialization for DCX and QZQSM messages with added unit tests and support modules.

* chore: ignore .plans directory in git configuration

* modified:   .gitignore

* fix: use acsesser in example ino files

* feat: implement Nankai Trough page aggregation logic and platform-independent time utilities

+++ src/Parser.h
@@ -40,7 +40,7 @@ private:
 
     // Process Nankai Trough page aggregation
     // Returns true if message should be output (page aggregation complete or not Nankai)
-    bool processNankaiAggregation(const Message& decoded, Message& out, uint32_t current_ms);
+    bool processNankaiAggregation(const Message& decoded, Message& out, uint64_t current_ms);
 };
 
 } // namespace azaraC
